refactor(web): extract the TTL cache helper out of the curated fallback
Moves withTtl into its own module and adds a keyed variant, so the
multichain clients can reuse it instead of copying the implementation.

The request-collapsing half of this helper is the part that matters: the
homepage renders every curated tab at once, and without collapsing, a cold
cache fires one upstream burst per tab. That is what made the keyless
CoinGecko tier answer 429 and silently cut the RWA tab from 11 priced rows
down to 3.
